# Use concrete types for path steps