--- a/cmp/compare.go
+++ b/cmp/compare.go
@@ -282,7 +282,7 @@ func (s *state) compareAny(step PathStep) {
 			return
 		}
 		vx, vy = vx.Elem(), vy.Elem()
-		s.compareAny(&indirect{pathStep{t.Elem(), vx, vy}})
+		s.compareAny(Indirect{&indirect{pathStep{t.Elem(), vx, vy}}})
 		return
 	case reflect.Interface:
 		if vx.IsNil() || vy.IsNil() {
@@ -294,7 +294,7 @@ func (s *state) compareAny(step PathStep) {
 			s.report(false, 0)
 			return
 		}
-		s.compareAny(&typeAssertion{pathStep{vx.Type(), vx, vy}})
+		s.compareAny(TypeAssertion{&typeAssertion{pathStep{vx.Type(), vx, vy}}})
 		return
 	default:
 		panic(fmt.Sprintf("%v kind not handled", t.Kind()))
@@ -322,7 +322,7 @@ func (s *state) tryMethod(t reflect.Type, vx, vy reflect.Value) bool {
 	return true
 }
 
-func (s *state) callTRFunc(f, v reflect.Value, step *transform) reflect.Value {
+func (s *state) callTRFunc(f, v reflect.Value, step Transform) reflect.Value {
 	v = sanitizeValue(v, f.Type().In(0))
 	if !s.dynChecker.Next() {
 		return f.Call([]reflect.Value{v})[0]
@@ -392,7 +392,7 @@ func sanitizeValue(v reflect.Value, t reflect.Type) reflect.Value {
 func (s *state) compareStruct(t reflect.Type, vx, vy reflect.Value) {
 	var vax, vay reflect.Value // Addressable versions of vx and vy
 
-	step := &structField{}
+	step := StructField{&structField{}}
 	for i := 0; i < t.NumField(); i++ {
 		step.typ = t.Field(i).Type
 		step.vx = vx.Field(i)
@@ -423,8 +423,8 @@ func (s *state) compareStruct(t reflect.Type, vx, vy reflect.Value) {
 }
 
 func (s *state) compareSlice(t reflect.Type, vx, vy reflect.Value) {
-	step := &sliceIndex{pathStep: pathStep{typ: t.Elem()}}
-	withIndexes := func(ix, iy int) *sliceIndex {
+	step := SliceIndex{&sliceIndex{pathStep: pathStep{typ: t.Elem()}}}
+	withIndexes := func(ix, iy int) SliceIndex {
 		if ix >= 0 {
 			step.vx, step.xkey = vx.Index(ix), ix
 		} else {
@@ -503,7 +503,7 @@ func (s *state) compareMap(t reflect.Type, vx, vy reflect.Value) {
 
 	// We combine and sort the two map keys so that we can perform the
 	// comparisons in a deterministic order.
-	step := &mapIndex{pathStep: pathStep{typ: t.Elem()}}
+	step := MapIndex{&mapIndex{pathStep: pathStep{typ: t.Elem()}}}
 	for _, k := range value.SortKeys(append(vx.MapKeys(), vy.MapKeys()...)) {
 		step.vx = vx.MapIndex(k)
 		step.vy = vy.MapIndex(k)